Split utils into models and utils files. Added class for transformers
